Moved std lib includes in TSS.CPP from stdafx.h to the library headers to eliminate necessity of their explicit inclusion in client app code. Fixed or suppressed remaining warnings.

+++ TSS.CPP/Samples/Samples.cpp
@@ -1726,7 +1726,7 @@ void Samples::Attestation()
                                                 nvHandle,
                                                 nonce,
                                                 TPMS_NULL_SIG_SCHEME(),
-                                                toWrite.size(),
+                                                (UINT16)toWrite.size(),
                                                 0);
 
     sigOk = pubKey.outPublic.ValidateCertifyNV(nonce, toWrite, 0, nvQuote);
